Split Result class for refactoring
